Reorganization of files, added handlers, and a little more.

+++ commands/restart.js
@@ -1,11 +1,19 @@
+require('dotenv').config();
+const TOKEN = process.env.TOKEN;
+
 module.exports = {
     name: 'restart',
     description: "Restarts Bot",
-    execute(message, args, client, TOKEN){
-        message.channel.send('Restarting...').then(m => {
-            client.destroy()
-            client.login(TOKEN);
-            message.channel.send('Restart Success');
-        }); 
+    execute(client, message, args, Discord, isBotOwner) {
+        if(!isBotOwner) {
+            message.channel.send("Error: Insufficient permission");
+            return;
+        } else {
+            message.channel.send('Restarting...').then(m => {
+                client.destroy()
+                client.login(TOKEN);
+                message.channel.send('Restart Success');
+            }); 
+        }
     }
 }
